gmi packages provides a Gemini text scanner
Add package gmi which currently provides a Gemini text scanner.  This
scanner is scans Gemini line by line using bufio.Scanner from the Go
standard library as the underlying scanner.

The scanner identifies which Gemini line type was scanned and provides
methods for extracting the token text from the line.  If the line was a
link, you can extract the text and the URL separately.

Scanning the file line by line like this has the advantage of being able
to scan long files and only allocate memory for the buffer to store a
single line of text at a time.

Benchmark tests are included and compare this implementation to the
bufio.Scanner and the Gemini parser: toast.cafe/x/gmi